+++ src/routes/search.routes.js
@@ -1,5 +1,5 @@
 import express from "express";
-import { test, isCorrectCafeId } from "../middlewares/cafeMiddleware.js";
+import { test, isCorrectCafeId } from "../middlewares/cafe-middleware.js";
 import { authenticateJWT } from "../middlewares/authMiddleware.js";
 import {
   cafeSearch,
